added pam_tty_conv function
this function is apparently intended to handle the communication
between the pam module and the user. The current function is a skeleton
that should take a pam message and send it to stdout.

Note: This has not been tested and is not currently called. This commit
allows the module to compile with this code, however.